% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card}
\alias{card}
\title{card}
\usage{
card(..., title = NULL, icon = shiny::icon("info"), footer = NULL,
  background = NULL, width = 3, height = NULL)
}
\arguments{
\item{...}{shiny elements}

\item{title}{card title}

\item{icon}{shiny icon}

\item{footer}{footer}

\item{background}{background}

\item{width}{in numer of columns; default 3}

\item{height}{px}
}
\value{
shiny element
}
\description{
A card like element for shiny.
}
\details{
Based on shinydashboard box element. Adding icons, shadows and a default width of 3.
}
\author{
Reinhard Simon
}
